Create Item.cs
For each item in the game, there will be the same class for it, so a ring and a chestplate will work the exact same. The ring would simply have a defense stat of 0 and have more effects, whereas the chestplate will have a higher defense stat but little to no effects. 

Same with weapons, the rings and armor simply have 0 attack stat, while the weapons have higher attack.